Clean typicode service and move implementation to resovlers
